# api_doc.yaml
openapi: 3.0.0
info:
  title: Documentación de la API
  description: Documentación de la API de Coches y Marcas
  version: 1.0.0

paths:
  /api/coches:
    get:
      summary: Obtener todos los coches
      responses:
        '200':
          description: Respuesta exitosa

    post:
      summary: Agregar un nuevo coche
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                nombre:
                  type: string
                fundacion:
                  type: string
                fundador:
                  type: string
                tipo:
                  type: string
                imagen:
                  type: string
              required:
                - nombre
                - fundacion
                - fundador
                - tipo
                - imagen
      responses:
        '201':
          description: Coche creado exitosamente
        '404':
          description: Error - No encontrado

  /api/coches/{id}:
    get:
      summary: Obtener un coche por su ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Respuesta exitosa
        '404':
          description: Error - No encontrado

    post:
      summary: Actualizar un coche por su ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                marca:
                  type: string
                modelo:
                  type: string
                version:
                  type: string
                año:
                  type: string
                combustible:
                  type: string
                imagen:
                  type: string
              required:
                - marca
                - modelo
                - version
                - año
                - combustible
                - imagen
      responses:
        '204':
          description: Coche actualizado exitosamente
        '404':
          description: Error - No encontrado

    delete:
      summary: Eliminar un coche por su ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Coche eliminado exitosamente
        '404':
          description: Error - No encontrado

  /api/marcas:
    get:
      summary: Obtener todas las marcas
      responses:
        '200':
          description: Respuesta exitosa

    post:
      summary: Agregar una nueva marca
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                fundacion:
                  type: string
                fundador:
                  type: string
                tipo:
                  type: string
                imagen:
                  type: string
              required:
                - nombre
                - fundacion
                - fundador
                - tipo
                - imagen
      responses:
        '201':
          description: Marca creada exitosamente
        '404':
          description: Error - No encontrado

  /api/marcas/{id}:
    get:
      summary: Obtener una marca por su ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Respuesta exitosa
        '404':
          description: Error - No encontrado

    post:
      summary: Actualizar una marca por su ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                fundacion:
                  type: string
                fundador:
                  type: string
                tipo:
                  type: string
                imagen:
                  type: string
              required:
                - nombre
                - fundacion
                - fundador
                - tipo
                - imagen
      responses:
        '204':
          description: Marca actualizada exitosamente
        '404':
          description: Error - No encontrado

    delete:
      summary: Eliminar una marca por su ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Marca eliminada exitosamente
        '404':
          description: Error - No encontrado